Move Autocomplete to separate file and add helper function
